sprint-5/step-2 made FeedOrderInfo for window
